Moved TextureGroup to texture module
